OH BOI. lotsa stuff added, changed and fixed and whatnot. basically, we now have working parameters. yay! and sum is fully implemented with that.
also, we can now keep the bot running instead of taking just one instruction.

the validation is also fixed.

the big change is the parameter stuff, as you can see there is a params field added to checkRule and handleRule, which is very important for this.

so far, so good.

and dont you ever say I dont do detailed commits.